requirements.txt
appdirs==1.4.3
astroid==2.1.0
attrs==18.2.0
backcall==0.1.0
black==18.9b0
bleach==3.1.0
calmap==0.0.7
Click==7.0
cycler==0.10.0
decorator==4.3.2
defusedxml==0.5.0
entrypoints==0.3
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
notebook==5.7.4
numpy==1.16.1
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
patsy==0.5.1
pexpect==4.6.0
pickleshare==0.7.5
prometheus-client==0.5.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pyarrow==0.12.0
Pygments==2.3.1
pylint==2.2.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.0.0
qtconsole==4.4.3
rise==5.4.1
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sns==0.1
statsmodels==0.9.0
terminado==0.8.1
testpath==0.4.2
toml==0.10.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.3.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
